faceRecog/requirements.txt
absl-py==0.9.0
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.2.7
astunparse==1.6.3
boto3==1.17.24
botocore==1.20.24
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.16
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.7.0
django-heroku==0.3.1
django-otp==1.0.2
django-storages==1.11.1
djangorestframework==3.12.2
drf-writable-nested==0.6.2
face-recognition-models==0.3.0
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gast==0.3.3
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.1
idna==2.10
image==1.5.33
importlib-metadata==1.7.0
importlib-resources==5.1.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mysql==0.0.2
mysqlclient==2.0.3
numpy==1.18.2
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.2.0.34
opt-einsum==3.3.0
payu-biz==2.0
Pillow==7.1.1
protobuf==3.12.4
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pyparsing==2.4.7
PyPDF2==1.26.0
PySide2==5.14.1
pytesseract==0.3.4
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
python-http-client==3.3.2
pytz==2020.1
razorpay==1.2.0
reportlab==3.5.64
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.4
scikit-learn==0.19.1
scipy==1.4.1
sendgrid==6.6.0
shiboken2==5.14.1
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tesseract==0.1.3
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.4.2
virtualenvwrapper-win==1.2.6
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.5
zipp==3.1.0